Fix endless RAF loop caused by web-view/progress.js
Progress.update was checking if new progress value was < 1 and increment progress or stop requesting frames and put progress to idle. Somehow on servo we still were getting extra update initiated by RAF in which case progress would be 0 we end up with restarted loop but this time it would never end since status remained idle and in that case we compute progress as 0.

This change makes sure that ticks on idle progress bar are ignored & there for no more new frames are requested.