Zoom in page on mousedown
This workaround #1038 and servo/servo#11664 that prevents us from cancelling click events when force touch is handled. With this change we react to mousedown event instead of click which fixes our issue since overlay is not displayed when zoomed in & there for we avoid causing zoom-in at the end of force touch. clicks.