Refactor `Input` to avoid mixins. (#1202)
* Make Editable and Focusable embeddable.

Editable and Focusable are implemented as interfaces or components that you need to mix in. This a step towards making them more standard components that can be embedded pretty much like all other components are.

* Refactor TextInput

Embed Focusable, Editable, Control instead of mixing them in.

* Refactor Input

Embed `Focusable`, `Editable` instead of mixing them in.